+++ World/Source/Scripts/Engines and Systems/Trades/Global Shoppe/Core/CustomerOrderShoppe.cs
@@ -1,0 +1,72 @@
+using System;
+using System.Collections.Generic;
+
+namespace Server.Engines.GlobalShoppe
+{
+    public abstract class CustomerOrderShoppe<TOrderContext> : CustomerShoppe, IOrderShoppe
+        where TOrderContext : IOrderContext
+    {
+        protected CustomerOrderShoppe(Serial serial) : base(serial)
+        {
+        }
+
+        protected CustomerOrderShoppe(int itemId) : base(itemId)
+        {
+        }
+
+        public void AddOrderItem(int index, Mobile from, TradeSkillContext context)
+        {
+            if (context.Orders.Count <= index) return;
+
+            var order = context.Orders[index];
+            if (order.IsComplete) return;
+
+            from.CloseGump(typeof(OrderGump));
+            from.SendGump(new OrderGump(from, order));
+        }
+
+        public void CompleteOrder(int index, Mobile from, TradeSkillContext context)
+        {
+            if (context.Orders.Count <= index) return;
+
+            var order = context.Orders[index];
+            if (!order.IsComplete) return;
+
+            context.Gold += order.GoldReward;
+            context.Points += order.PointReward;
+            context.Reputation = Math.Min(ShoppeConstants.MAX_REPUTATION, context.Reputation + order.ReputationReward);
+            context.Orders.Remove(order);
+
+            from.PlaySound(0x32); // Dropgem1
+        }
+
+        public void RejectOrder(int index, TradeSkillContext context)
+        {
+            if (context.Orders.Count <= index) return;
+
+            var order = context.Orders[index];
+            context.Reputation = Math.Max(0, context.Reputation - order.ReputationReward);
+
+            context.Orders.Remove(order);
+        }
+
+        protected abstract IEnumerable<TOrderContext> CreateOrders(Mobile from, TradeSkillContext context, int amount);
+
+        protected override void OpenGump(Mobile from, TradeSkillContext context)
+        {
+            if (context.FeePaid && context.CanRefreshOrders)
+            {
+                var count = ShoppeConstants.MAX_ORDERS - context.Orders.Count;
+                foreach (var order in CreateOrders(from, context, count))
+                {
+                    context.Orders.Add(order);
+                }
+
+                context.CanRefreshOrders = false;
+                context.NextOrderRefresh = DateTime.UtcNow.Add(ShoppeConstants.ORDER_REFRESH_DELAY);
+            }
+
+            base.OpenGump(from, context);
+        }
+    }
+}
