+++ World/Source/Scripts/Engines and Systems/Trades/Global Shoppe/Core/IOrderShoppe.cs
@@ -1,0 +1,11 @@
+namespace Server.Engines.GlobalShoppe
+{
+    public interface IOrderShoppe
+    {
+        void AddOrderItem(int index, Mobile from, TradeSkillContext context);
+
+        void CompleteOrder(int index, Mobile from, TradeSkillContext context);
+
+        void RejectOrder(int index, TradeSkillContext context);
+    }
+}
